Added results window
Results window still needs to be worked on to show results, but it will pop up after the user clicks the calculate button if the display option is checked. So far the results window only shows headers for each operation selected and headers for pre and post test results. Also needs an area for graphs to be displayed.